Extract parametric type metaclass from FunsorMeta and domains
